Size of Advisor Widget and Updating Advisor Details
Changed the column specification so the left column would be smaller in width and chanegd the css of the pictures so they are smaller. I could not find how to change the descriptions however.

I also changed the advisor details so the state fo teh advisor and their profile would be stored and updated when changing between advisors. It takes a long time to load between changing advisors but I can look into that.